src/quality_verification_layer/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quality-verification-layer"
version = "0.1.0"
description = "Agnes Quality Verification Layer - evidence-based supplier quality verification"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0,<3.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "pdfplumber>=0.10",
    "google-genai>=1.0",
]

[tool.setuptools.packages.find]
include = ["quality_verification_layer*"]

[project.scripts]
quality-verification = "quality_verification_layer.cli:main"
